Fixed how we handle unions of records with the same fields.
When deserializing an Avro object, we need to select records based on their names, not just their fields. This lets us have unions with two records that have different names but the same set of fields (which is totally valid in Avro).

In principle, we can select named types based *only* on names, not looking at the fields at all. Avro requires full names in schemas to be unique. However, I did not make this change here to avoid changing the existing behavior of the library too much.