Fixed how default values are parsed for bytes and fixed types.
Default values for bytes objects are encoded as JSON string literals where each codepoint is between 0 and 255 and directly represents a byte.

The biggest change here actually involves Template Haskell. The previous deriving code generated schemas by encoding the schema to JSON at compile time, embedding it as a string literal and parsing it at runtime. This caused problems with the encoding of bytes and generally seemed error-prone, so I replaced it with a whole bunch of TH code to generate the Schema object directly.