Properly handle namespaces according to Avro spec.
This PR ensures that we properly infer and handle namespaces in Avro specs. In code, this is reflected by adding a `namespace` field to `TypeName` and taking the namespace field *out* of `Record`, `Enum` and `Fixed`:

```
data TypeName = TN { baseName  :: T.Text
                   , namespace :: [T.Text]
                   }
```

This ensures that we /always/ consider namespaces when comparing Avro type names, using them as keys... etc. The idea is that we take care of inferring and normalizing namespaces in parsing so that 'TypeName' values are 100% canonical in the rest of the code—we can consistently compare them for equality regardless of how they were specified in the Avro schema itself.

As a bonus, this let us simplify a few awkward edgecases in the code that tried to deal with namespace logic. Every part of the code outside of parsing now uses `TypeName`s as opaque values without caring about names vs namespaces.

The actual namespace inference logic is handled in the `FromJSON` instance for `Schema` and follows the rules laid out in the Avro specification. Given a definition of a named type (Record, Enum or Fixed) with a given `"name"` and `"namespace"` field, the fullname is defined as follows:

  1. If the `"name"` field contains dots, it is parsed as a /fullname/ and the `"namespace"` field is ignored if present. The name field `"com.example.Foo"` would be parsed as a type name with the name `"Foo"` and the namespace `"com.example"`.

  2. Otherwise, if both the `"name"` and `"namespace"` fields are set, they make up the /fullname/.

  3. If only the `"name"` field is specified, the `"namespace"` is inferred from the namespace of the /most tightly enclosing schema/, if any. If there is no enclosing schema, the namespace is null. (In `TypeName`, this corresponds to a namespace of `[]`.)

The namespace logic for references to named types follows steps 1 and 3.

This fixes all the problems I brought up in #19:

  * there is no distinction between `Nothing` and `Just ""`—both become `[]` in the new scheme
  * `TypeName` now contains both the name and the namespace, forcing us to handle namespaces consistently throughout the codebase
  * namespaces are inferred following the rules laid out in the Avro specification (which weren't as ugly as I though)

This PR also updates the Template Haskell code to handle namespaces correctly. A type like `com.example.Foo` becomes the Haskell type `Com'example'Foo`.

However, this breaks compatibility with the old Template Haskell code and leads to *very* verbose names if your schemas use non-trivial namespaces. To maintain backwards compatibility and let people have nicer generated type names, the Template Haskell code can be configured to ignore namespaces. Now there are four functions to derive Avro types:

  * `deriveAvro` which ignores namespaces and derives to/from Avro instances
  * `deriveAvroWithNamespaces` which handles namespaces and derives to/from Avro instances
  * `deriveFromAvro` which ignores namespaces and derives only from Avro instances
  * `deriveFromAvroWithNamespaces` which handles namespaces and derives only from Avro instances