Fixed Data.Avro.Schema.extractBindings.
The old implementation had some odd bugs with nested type definitions, including missing types defined inside arrays and maps. This PR has a new implementation of extractBindings that fixes all the problems as well as a test case which covers a bunch of edge cases for extracting types from a schema.