Fixed how we decode JSON unions with no namespaces.
The old logic would error out if the JSON encoding used a name like `Foo` while the schema used `.Foo` (or vice-versa). `Foo` and `.Foo` should be treated equivalently as "`Foo` with no namespace".

This PR adds an explicit canonicalization step in the JSON logic to take care of this problem. This step required a bit of extra bookkeeping to deal with primitive types like `string` so that we don't try using `.string`.

I also added a regression test to `JSONSpec.hs` to catch this problem in the future.