Fixed parsing unions with arrays and maps in JSON.
An Avro union can contain up to one array and one map type as options. In JSON, a value of a union that is an array is represented as follows:

```
{ "array" : [1,2,3] }
```

and maps are

```
{ "map" : { "a" : 1, "b" : 2 } }
```

Currently, trying to parse an object like this with an appropriate schema gives us the following error:

```
Error "Type 'array' not in union: Array {item = String} :| [Map {values = Long}]"
```

After this PR, these objects will be parsed correctly according to the schema.