Fixed the implementation of getMap.
The previous version of getMap did not handle blocks that began with a negative number. (See #85 for details.) This PR fixes the behavior and shares the code to handle blocks between `getMap` and `getArray`.

I'm pretty sure we need a similar fix in `Decode.Lazy` and `Decode.Strict`, but I don't fully understand what's going on in those modules and don't really want to touch them right now.